staff-core/src/main/java/com/btobits/automator/fix/ant/filter/comparators/EqualsComparator.java
/*
  * Copyright © 2011-2014 EPAM Systems/B2BITS® (http://www.b2bits.com).
 *
 * This file is part of STAFF.
 *
 * STAFF is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Lesser General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * STAFF is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 * GNU Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public License
 * along with STAFF. If not, see <http://www.gnu.org/licenses/>.
 */

package com.btobits.automator.fix.ant.filter.comparators;

/**
 * @author Mykhailo_Sereda
 */
public class EqualsComparator extends AbstractFilterComparator {
    
    @Override
    public boolean compare(String value1, String value2) {
        validateNullValues(value1, value2);
        if (value1.compareTo(value2) == 0) {
            return true;
        } else {
            return false;
        }
    }
}
